+++ qgrepControls/Properties/Settings.Designer.cs
@@ -406,5 +406,29 @@ namespace qgrepControls.Properties {
                 this["WindowSizes"] = value;
             }
         }
+        
+        [global::System.Configuration.UserScopedSettingAttribute()]
+        [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
+        [global::System.Configuration.DefaultSettingValueAttribute("False")]
+        public bool CppHeaderInclusion {
+            get {
+                return ((bool)(this["CppHeaderInclusion"]));
+            }
+            set {
+                this["CppHeaderInclusion"] = value;
+            }
+        }
+        
+        [global::System.Configuration.UserScopedSettingAttribute()]
+        [global::System.Diagnostics.DebuggerNonUserCodeAttribute()]
+        [global::System.Configuration.DefaultSettingValueAttribute("[]")]
+        public string History {
+            get {
+                return ((string)(this["History"]));
+            }
+            set {
+                this["History"] = value;
+            }
+        }
     }
 }